% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{allelematch::amExample1}
\alias{allelematch::amExample1}
\title{This is sample data from package allelematch that we use to test
allelmatch backwards compatibility.}
\description{
This is sample data from package allelematch that we use to test
allelmatch backwards compatibility.
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
