% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{amExample1_0100_expected}
\alias{amExample1_0100_expected}
\alias{amExample1_0101_expected}
\alias{amExample1_0102_example1_1_expected}
\alias{amExample1_0103_example1_2_expected}
\alias{amExample1_0104_expected}
\alias{amExample1_0105_expected}
\alias{amExample2_0100_expected}
\alias{amExample2_0101_expected}
\alias{amExample2_0102_example2_1_expected}
\alias{amExample2_0103_expected}
\alias{amExample2_0104_example2_2_expected}
\alias{amExample2_0105_expected}
\alias{amExample2_0106_expected}
\title{Expected data used in calls to \link[testthat:equality-expectations]{testthat::expect_identical}}
\description{
Naming convention: "exp" for "Expected". First number identifies the \link[testthat:test_that]{testthat::test_that}
code block. The second number identifies the use within the same code block. \cr\cr
}
\details{
This expected data was generated from runing the tests towards version 2.5.3 of \link{allelematch}.
Look for commented out lines ending with "# This is how the expected data was stored" to see how.
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
