% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{amExample1_inp}
\alias{amExample1_inp}
\title{This is sample data from package allelematch version 2.5.1,
but saved on .csv format, as described under 'Details' in utils::data.}
\description{
as described  that we use to test
allelmatch backwards compatibility.
... saved on the format
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
