% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{amUniqueWrapper}
\alias{amUniqueWrapper}
\title{Wrapper around \link[allelematch:amUnique]{allelematch::amUnique} that adds logging}
\usage{
amUniqueWrapper(amDatasetFocal, ...)
}
\description{
Generates a message string that describes the specified parameters.\cr
Feeds the specified parameters to \link[allelematch:amUnique]{allelematch::amUnique}.\cr
}
