% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artAssertAllelematchVersion}
\alias{artAssertAllelematchVersion}
\title{Makes sure further testing uses \code{wantedVersion} of package allelematch}
\usage{
artAssertAllelematchVersion(wantedVersion = c("2.5.3", "2.5.2"))
}
\arguments{
\item{wantedVersion}{a string that contains the wanted version of allelematch.
Note that the current version of R,  4.3.2, no longer accepts allelematch "2.5.1".
The first supported version is "2.5.2".}
}
\value{
The wantedVersion on success, otherwise execution is stopped.
}
\description{
If the \code{wantedVersion} of package \code{allelematch} is not yet installed,
then the current version is deleted, the wantedVersion is installed and
R is restarted.
}
