% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{artAssertLocalDataIdentical}
\alias{artAssertLocalDataIdentical}
\title{Stops execution if remote and local data are not identical}
\usage{
artAssertLocalDataIdentical(package, remoteName, localName = remoteName)
}
\arguments{
\item{package}{The name of the remote package to load from as character}

\item{remoteName}{The name of the data to read from the remote package}

\item{localName}{The name of the data to write to the local data dir as .R file}
}
\description{
Stops execution if remote and local data are not identical
}
