% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_alparam.R
\name{artDefaultFileName}
\alias{artDefaultFileName}
\title{Turns \link{artArgToString} into a string that can be used in a filename}
\usage{
artDefaultFileName(...)
}
\arguments{
\item{...}{The \code{allelematch} parameters to abbreviate into a file name.}
}
\value{
a string that can be used in a data file name for use in an allelematch regressiontest.
}
\description{
Here we assume that the arguments in \link{artArgToString} are those
passed on to an \link{allelematch} function by a \link{amregtest} wrapper function. \cr\cr
}
\details{
The filenames are then used to store data for expected and actual output
from tests.
}
\keyword{internal}
