% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artInstallAllelematchVersion}
\alias{artInstallAllelematchVersion}
\title{Installs the version of package \code{allelematch} to test}
\usage{
artInstallAllelematchVersion(wantedVersion)
}
\arguments{
\item{wantedVersion}{a string that contains the wanted version of \code{allelematch}.
Note that the current version of R,  4.3.2, no longer accepts \code{allelematch} "2.5.1".
The first supported version is "2.5.2".}
}
\value{
The wantedVersion on success, otherwise execution is stopped.
}
\description{
If the \code{wantedVersion} of package \link{allelematch} is not yet installed,
then the current version is deleted, the \code{wantedVersion} is installed and
R is restarted.\cr
\cr
Look at \href{https://github.com/cran/allelematch/tags}{allelematch tags} at CRAN to see available versions.
\cr
Note that the first supported version is 2.5.2. Older versions are no longer
tolerated modern versions of R (e.g. 4.3.2).

TODO: Find the change and revision in R!
}
\seealso{
\link{artVersion} \link{artAssertAllelematchVersion}
}
