% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_api.R
\name{artList}
\alias{artList}
\title{Lists available tests in \code{amregtest} without running them}
\usage{
artList(verbose = TRUE)
}
\arguments{
\item{verbose}{logical. If TRUE (the default), prints additional info to stdout}
}
\value{
A character vector containing the names of all the tests
}
\description{
Use the output to select a value for parameter \code{filter} to \link{artRun}.
Useful when debugging.
}
\examples{
# See what version of packages 'allelematch' and 'amregtest'
# are currently loaded:
artVersion()

# List the available tests:
artList()
\donttest{
# Run all the tests:
# artRun()  # Takes several minutes

# Run the first of the available tests:
artRun(filter="allelematch_1-amDataset$")
}

}
\seealso{
\link{artVersion} and \link{artRun}
}
