% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_alin.R
\name{artLoadDataSet}
\alias{artLoadDataSet}
\title{Loads an loggs the \link[allelematch:amDataset]{allelematch::amDataset}}
\usage{
artLoadDataSet(dataName)
}
\arguments{
\item{dataName}{Name of the data set to load from ./data/.}
}
\description{
Loads an loggs the \link[allelematch:amDataset]{allelematch::amDataset}
}
\keyword{internal}
