% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{artOverwriteExpected}
\alias{artOverwriteExpected}
\title{Write data to .R file in ./data directory using \link[base:dump]{base::dump}`}
\usage{
artOverwriteExpected(df, outName, overwrite = FALSE)
}
\arguments{
\item{df}{Data frame to write}

\item{outName}{Name of file to write to, excluding leading "data/" and trailing ".R"}

\item{overwrite}{Set to TRUE to write df as new expected data to dir "data/.".
i.e. ./data under the package source root directory.}
}
\description{
Used when creating new tests that need new expected data.
}
