% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_text.R
\name{artReadCsvFile}
\alias{artReadCsvFile}
\title{Read gg style .csv file for use in ... (\link[allelematch:amDataset]{allelematch::amDataset} ?)}
\usage{
artReadCsvFile(csvInFile)
}
\arguments{
\item{csvInFile}{Dir, name and extension for the .csv file to be read}
}
\description{
The format of the data to be analyzed by \code{allelematch}
is outside of the control by \code{allelematch}. So, the format read by this function
is specific to this package rather than to \code{allelematch}.
}
\keyword{internal}
