% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_api.R
\name{artRefreshAllelematch}
\alias{artRefreshAllelematch}
\title{Ensures that latest installed 'allelematch' is also loaded in this R session.}
\usage{
artRefreshAllelematch(verbose = FALSE)
}
\arguments{
\item{verbose}{logical. If TRUE (the default), prints diagnostic progress
messages to stdout}
}
\value{
TBD
}
\description{
Ensures that the latest installed version of 'allelematch' is loaded and
attached in this R session, regardless of version string.
Any older version is unloaded.

This function handles the case where 'allelematch' development and 'amregtest'
testing of 'allelematch' is done in different RStudio sessions.
Development implies frequent repetitive edit-build-install cycles
without changing the 'allelematch' version. These re-installations
are not automatically detected by R in another R session, causing large
risk of testing an older version of 'allelematch' than intended.

This function is also called from \link{artRun}, \link{artVersion},
and from the setup.R file in the tests/testthat directory.
}
\examples{
# Install the default official version of 'allelematch' from CRAN:
artInstallCranAllelematch()

# Install another official version of 'allelematch' from CRAN:
artInstallCranAllelematch("2.5.3")

}
\seealso{
\link{artVersion}, \link{artList}, \link{artRun} and \link{amregtest}
}
