% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artSet}
\alias{artSet}
\title{Controls execution of the regressiontest}
\usage{
artSet(skip_slow = NA, generate_html = NA)
}
\arguments{
\item{skip_slow}{TRUE or FALSE}

\item{generate_html}{TRUE or FALSE}
}
\description{
Two envionement variables control the executon of the tests
then the current version is deleted, the \code{wantedVersion} is installed and
R is restarted.\cr
\cr
Look at \href{https://github.com/cran/allelematch/tags}{allelematch tags} at CRAN to see available versions.
\cr
Note that the first supported version is 2.5.2. Older versions are no longer
tollerated modern versions of R (e.g. 4.3.2).

TODO: Find the change and revision in R!
}
\seealso{
\link{artShow} \link{artClear}
}
