% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{artVerifyAmExamples}
\alias{artVerifyAmExamples}
\title{Compares the local amExample files with the originals.}
\usage{
artVerifyAmExamples()
}
\description{
Stops execution if any of the local amExample files is not identical with the corresponding
file in the \link{allelematch} package
}
