% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_text.R
\name{artWriteCsv2File}
\alias{artWriteCsv2File}
\title{Write .csv file on format accepted by \link[utils:data]{utils::data} and excel.}
\usage{
artWriteCsv2File(df, csvOutFile)
}
\arguments{
\item{df}{Data to be written}

\item{csvOutFile}{Name of file to write, or an already open \link{connection}. \cr
An open connection will be left open. \cr
A named file will be closed at exit. \cr}
}
\description{
Write .csv file on format accepted by \link[utils:data]{utils::data} and excel.
}
