% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_imports.R
\name{au_imports}
\alias{au_imports}
\title{Placeholder function for roxygen2 import tags}
\usage{
au_imports()
}
\description{
This package uses the recommended "Single file" strategy from the roxygen2
documentation for how to maintain import dependencies,
see ????
}
\details{
Roxygen uses the tags in our R/*.R files to automatically update the NAMESPACE
file.

See https://cran.r-project.org/doc/manuals/R-exts.html#Package-Dependencies
for how to manually add needed packages to the 'Imports' 'Depends', 'Suggests'
or 'Enhances' statements in the DESCRIPTION file.

Hint: Use \code{pak::pkg_deps_tree(package)} (e.g. pak::pkg_deps_tree("dplyr"))
to see the package dependencies.
}
