% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_alparam.R
\name{auArgToString}
\alias{auArgToString}
\title{Turns \code{...} into a string of \if{html}{\out{<name>}}=\if{html}{\out{<value>}} pairs}
\usage{
auArgToString(...)
}
\arguments{
\item{...}{The arguments to be converted to text}
}
\value{
a string of \if{html}{\out{<name>}}=\if{html}{\out{<value>}} pairs
}
\description{
The \if{html}{\out{<name>}}=\if{html}{\out{<value>}} pairs are separated by ", ".\cr\cr
}
\details{
For use when logging the input to a function that takes \code{...} parameters.
}
