% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{auAssertAllelematchVersion}
\alias{auAssertAllelematchVersion}
\title{Makes sure further testing uses \code{wantedVersion} of package allelematch}
\usage{
auAssertAllelematchVersion(wantedVersion = c("2.5.3", "2.5.2"))
}
\description{
If the \code{wantedVersion} of package \code{allelematch} is not yet installed,
then the current version is deleted, the wantedVersion is installed and
R is restarted.
}
