% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{auAssertLocalDataIdentical}
\alias{auAssertLocalDataIdentical}
\title{Stops execution if remote and local data are not identical}
\usage{
auAssertLocalDataIdentical(package, remoteName, localName, keep = FALSE)
}
\arguments{
\item{package}{The name of the remote package to load from as character}

\item{remoteName}{The name of the data to read from the remote package}

\item{localName}{The name of the data to write to the local data dir as .R file}

\item{keep}{If TRUE, both remoteName and localName will be kept in the the .GlobalEnv. Otherwise they will be removed.}
}
\description{
Stops execution if remote and local data are not identical
}
