% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_alin.R
\name{auCheckCleanSamplesDf}
\alias{auCheckCleanSamplesDf}
\title{Checks that a data set has the expected GG format}
\usage{
auCheckCleanSamplesDf(cleanSamples)
}
\arguments{
\item{cleanSamples}{The data frame to be checked}
}
\description{
(Move out of this file? Too GG specific?)
}
