% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{auCopyAmExamplesToLocalData}
\alias{auCopyAmExamplesToLocalData}
\title{Copy from package \code{alleledata} all amExample?.RData files to this data directory.}
\usage{
auCopyAmExamplesToLocalData()
}
\description{
Copy from package \code{alleledata} all amExample?.RData files to this data directory.
}
