% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_alparam.R
\name{auDefaultFileName}
\alias{auDefaultFileName}
\title{Turns \link{auArgToString} into a string that can be used in a filename}
\usage{
auDefaultFileName(...)
}
\arguments{
\item{...}{}
}
\value{
a string that can be used in a data filename for use in an allelematch regressiontest.
}
\description{
Here we assume that the arguments in \link{auArgToString} are those
passed on to an \link{allelematch} function by a \link{regtest} wrapper function. \cr\cr
}
\details{
The filenames are then used to store data for expected and actual output
from tests.
}
