% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{auDumpToData}
\alias{auDumpToData}
\title{Write data to .R file in ./data directory using \code{dump()}}
\usage{
auDumpToData(df, outName)
}
\arguments{
\item{df}{Data frame to write}

\item{outName}{Name of file to write to, excluding leading "data/" and trailing ".R"}
}
\description{
Write data to .R file in ./data directory using \code{dump()}
}
