% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{auImportAmExamplesToLocalData}
\alias{auImportAmExamplesToLocalData}
\title{Import all \link[allelematch:amExampleData]{allelematch::amExampleData} data to this data directory.}
\usage{
auImportAmExamplesToLocalData(dir = NULL)
}
\arguments{
\item{dir}{A directory that contains the files (with extension ".RData") to import.
If set, installed package \code{allelematch} is ignored}
}
\description{
Import from package \link{allelematch} (or from specified \code{dir})
all \link[allelematch:amExampleData]{allelematch::amExampleData} .RData files to this data directory.
}
