% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{auImportToLocalData}
\alias{auImportToLocalData}
\title{Load data from remote package and dump it to .R file in local data directory}
\usage{
auImportToLocalData(package, remoteName, localName = remoteName, dir = NULL)
}
\arguments{
\item{package}{The name of the package to load from as character. Ignored if \code{dir} is set}

\item{remoteName}{The name of the data to read from the remote package}

\item{localName}{The name of the data to write to the local data dir as .R file. Default is \code{remoteName}}

\item{dir}{A directory that contains \code{remoteName}, with extension ".RData". If set, \code{package} is ignored}
}
\description{
Load data from remote package and dump it to .R file in local data directory
}
