% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{auObsoleteOutputData}
\alias{auObsoleteOutputData}
\alias{output_aMm15_actual.brief}
\alias{output_aMm15_actual}
\alias{output_aMm15_actual.sorted}
\alias{output_mThr0.9_actual.brief}
\alias{output_mThr0.9_actual}
\alias{output_mThr0.9_actual.sorted}
\title{To be removed.}
\description{
To be removed.
}
\keyword{data}
