% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{auOverwriteExpected}
\alias{auOverwriteExpected}
\title{Write data to .R file in ./data directory using \link[base:dump]{base::dump}`}
\usage{
auOverwriteExpected(df, outName, dir, overwrite = FALSE)
}
\arguments{
\item{df}{Data frame to write}

\item{outName}{Name of file to write to, excluding leading "data/" and trailing ".R"}

\item{dir}{The directory in which to write the data file.
Expected to point at the here::here("data") directory,
i.e. ./data under the package source root directory.}

\item{overwrite}{Set to ´TRUE´ for the overwrite to take place}
}
\description{
Write data to .R file in ./data directory using \link[base:dump]{base::dump}`
}
