% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_text.R
\name{auRead_amCSV}
\alias{auRead_amCSV}
\title{Read .csv file written by \link{allelematch} for comparison with data}
\usage{
auRead_amCSV(csvInFile)
}
\arguments{
\item{csvInFile}{Dir, name and extension for the .csv file to be read}
}
\description{
\link{allelematch} writes summary file using "," as field delimiter, rather
than ";", expected by \link[utils:data]{utils::data}.
\cr
Quote from \code{allelematch::amCSV.amUnique}  \cr
\cr
\code{utils::write.csv(csvTable, file=csvFile, row.names=FALSE)}
}
