% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{auVerifyAmExamples}
\alias{auVerifyAmExamples}
\title{Compares the local amExample files with the originals.}
\usage{
auVerifyAmExamples()
}
\description{
Stops execution if any of the local amExample files is not identical with the corresponding
file in the \code{allelematch} package
}
\examples{
auVerifyAmExamples()
}
