% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_text.R
\name{auWrite_amCSV}
\alias{auWrite_amCSV}
\title{Write .csv file on the same format as \link{allelematch}}
\usage{
auWrite_amCSV(df, csvOutFile)
}
\arguments{
\item{df}{Data to be written on \link{allelematch} summary format}

\item{csvOutFile}{The file to write to}
}
\description{
\link{allelematch} writes summary file using "," as field delimiter, rather
than ";", expected by \link[utils:data]{utils::data}.
}
\details{
Quote from \code{allelematch::amCSV.amUnique}  \cr\cr

\code{utils::write.csv(csvTable, file=csvFile, row.names=FALSE)}
}
