% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_text.R
\name{auWriteCsvFile}
\alias{auWriteCsvFile}
\title{Write .csv file on the same format as \link{allelematch}}
\usage{
auWriteCsvFile(df, csvOutFile)
}
\description{
\link{allelematch} writes summary file using "," as field delimiter, rather
than ";", expected by \link[utils:data]{utils::data}.
}
\details{
Quote from \code{allelematch::amCSV.amUnique}  \cr\cr

\code{utils::write.csv(csvTable, file=csvFile, row.names=FALSE)}
}
