% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{exp_01_00}
\alias{exp_01_00}
\alias{exp_01_01}
\alias{exp_01_02}
\alias{exp_01_02_example1_1}
\alias{exp_01_03_example1_2}
\alias{exp_01_04}
\alias{exp_01_05}
\alias{exp_02_00}
\alias{exp_02_01}
\alias{exp_02_02_example2_1}
\alias{exp_02_03}
\alias{exp_02_04_example2_2}
\alias{exp_02_05}
\alias{exp_02_06}
\title{Expected data used in calls to \link[testthat:equality-expectations]{testthat::expect_identical}}
\description{
Naming convention: "exp" for "Expected". First number identifies the \link[testthat:test_that]{testthat::test_that}
code block. The second number identifies the use within the same code block. \cr\cr
}
\details{
This expected data was generated from runing the tests towards version 2.5.3 of \link{allelematch}.
Look for commented out lines ending with "# This is how the expected data was stored" to see how.
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
