% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{getdata}
\alias{getdata}
\title{Loads and returns a large data set without polluting the global environment.}
\usage{
getdata(...)
}
\arguments{
\item{...}{Passed on to \code{utils::data} as is}
}
\value{
the specified data set loaded by \code{utils::data}
}
\description{
The data set is taken from the ./data/ folder of the specified package.
See \link{data} for how the supported data types are encrypted.
}
\details{
The parameters are the same as for the \code{utils::data()} function.

Thank you @henfiber! See https://stackoverflow.com/questions/30951204/load-dataset-from-r-package-using-data-assign-it-directly-to-a-variable/30951700#30951700
}
