% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{ggExampleData}
\alias{ggExampleData}
\alias{input_Match_references_clean}
\alias{input_new_samples_clean}
\alias{input_new_samples_clean_split}
\alias{output_aMm15_expected}
\alias{output_mThr0.9_expected}
\title{Data sets originating from GG work}
\description{
Large data sets gathered from field work in 2022.
\tabular{clcl}{
\verb{  } \tab \code{input_Match_references_clean  }\tab \verb{  } \tab A reference db containing known individuals\cr
\verb{  } \tab \code{input_new_samples_clean}\tab \tab A set of new samples from the field to be matched against the reference\cr
\verb{  } \tab \code{input_new_samples_clean_split  }\tab \tab The new samples cleaned up and alligned with the reference\cr
\verb{  } \tab \code{output_aMm15_expected}\tab \tab Output after running through \verb{amUnique( alleleMisMatch=15 (??))}...\cr
\verb{  } \tab \code{output_mThr0.9_expected}\tab \tab Output after running through ...
}
New attempt:

\tabular{cll}{
\verb{  } \tab \code{input_Match_references_clean}\tab A reference db containing known individuals\cr
\verb{  } \tab \code{input_new_samples_clean}\tab A set of new samples from the field to be matched against the reference\cr
\verb{  } \tab \verb{input_new_samples_clean_split \\ \\ \\ \\. }\tab The new samples cleaned up and alligned with the reference\cr
\verb{  } \tab \code{output_aMm15_expected}\tab Output after running through \verb{amUnique( alleleMisMatch=15 (??))}...\cr
\verb{  } \tab \code{output_mThr0.9_expected}\tab Output after running through ...
}
}
\keyword{data}
