% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_data.R
\docType{data}
\name{ggSample}
\alias{ggSample}
\title{Data sets originating from GG work}
\format{
This data is saved on semicolon (";") delimited .csv format, as described under 'Details' in \link[utils:data]{utils::data}.\cr
Data samples in rows, and alleles in columns. Missing data is represented as "-99".
}
\description{
Large data set gathered from field work in 2022. Here used to test package \link[allelematch:allelematch-package]{allelematch} for backwards compatibility.
Combines a reference db of known individuals with new samples to be analyzed.\cr\cr
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
\keyword{internal}
