% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiontest-package.R
\docType{package}
\name{regressiontest-package}
\alias{regressiontest}
\alias{regressiontest-package}
\title{regressiontest: Used to test package allelematch}
\description{
This package is used to automate regression testing when testing improvments of package allelematch.
}
\author{
\strong{Maintainer}: Torvald Staxler \email{torvald.staxler@gmail.com}

}
\keyword{internal}
