% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{summary_amUniqueWrapper}
\alias{summary_amUniqueWrapper}
\title{Wrapper around \link[allelematch:amUnique]{allelematch::summary.amUnique} that adds logging}
\usage{
summary_amUniqueWrapper(
  amUniqueOutput,
  outputData = NULL,
  expectedData,
  summary = FALSE,
  overwriteExpected = FALSE
)
}
\description{
Generates a message string that describes the specified parameters.\cr
Feeds the specified parameters to \link[allelematch:amUnique]{allelematch::summary.amUnique}.\cr
}
