% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_data.R
\docType{data}
\name{amExample1}
\alias{amExample1}
\title{Example 1 High quality data set}
\format{
Data frame with samples in rows, and alleles in columns. Missing data is represented as "-99".
}
\description{
This is sample data copied from \link[allelematch:amExampleData]{amExampleData} in version 5.2.1
of package \link[allelematch:allelematch-package]{allelematch}. We use this data to test 'allelematch' backwards compatibility.
}
\details{
The data in this example is simulated to represent a high quality data set that might
result from a laboratory protocol where samples were run multiple times to confirm their
identity. It has no genotyping error, a near-zero missing data load, and approximately
60\% of the individuals have been artificially resampled more than once.
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
