% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_data.R
\docType{data}
\name{amExample2}
\alias{amExample2}
\title{Example 2 Good quality data set}
\format{
Data frame with samples in rows, and alleles in columns. Missing data is represented as "-99".
}
\description{
This is sample data copied from \link[allelematch:amExampleData]{amExampleData} in version 5.2.1
of package \link[allelematch:allelematch-package]{allelematch}.
}
\details{
The data in this example have also been simulated, this time to reflect the qualities
of good quality data set, where genotyping error and missing data exist, but these can
be confidently handled by allelematch without manual intervention. At each locus a
random 4\% of heterozygotes lost their second allele to simulate an allele dropout, and
a random 4\% of samples at each locus had alleles set to missing.
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
