% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_data.R
\docType{data}
\name{amExample3}
\alias{amExample3}
\title{Example 3 Marginal quality data set}
\format{
Data frame with samples in rows, and alleles in columns. Missing data is represented as "-99".
}
\description{
This is sample data copied from \link[allelematch:amExampleData]{amExampleData} in version 5.2.1
of package \link[allelematch:allelematch-package]{allelematch}.
}
\details{
The data in this example have been simulated to represent a data set of marginal
quality where the use of allelematch combined with careful manual review of the
results is required to achieve a confident assessment of the unique genotypes. At each
locus a random 4\% of heterozygotes lost their second allele to simulate an allele dropout,
and a random 10\% of samples at each locus had alleles set to missing.
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
