% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{amExample5}
\alias{amExample5}
\title{Example 5}
\description{
This is sample data copied from \link[allelematch:amExampleData]{allelematch::amExampleData} in version 5.2.1
of package \link{allelematch}. We use this data to test allelmatch backwards compatibility.
}
\references{
\url{https://github.com/cran/allelematch}
}
\keyword{data}
