% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\docType{data}
\name{amExampleData}
\alias{amExampleData}
\title{Example data from package allelematch}
\description{
This example data is used when testing allelmatch backwards compatibility.\cr
It was imported from version 5.2.1 of \link{allelematch}.
It is still unchanged in 5.2.3. \cr
\cr
See \link[allelematch:amExampleData]{allelematch::amExampleData} for a brief introduction. \cr
See \href{https://github.com/cran/allelematch/blob/2.5.1/inst/doc/allelematchSuppDoc.pdf}{allelematchSuppDoc.pdf}.
for a more detailed description.
}
\examples{
# Make sure that the local copy of all data/amExample* files match
# match those in the currently installed version of allelematch:
artVerifyAmExamples() # Stops if not identical

}
\references{
\url{https://github.com/cran/allelematch}

\href{https://github.com/cran/allelematch/blob/2.5.1/inst/doc/allelematchSuppDoc.pdf}{allelematchSuppDoc.pdf}
}
\keyword{data}
