% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amregtest-package.R
\docType{package}
\name{amregtest-package}
\alias{amregtest-package}
\title{amregtest: Runs Allelematch Regression Tests}
\description{
Automates regression testing of package 'allelematch'. Over 2500 tests cover all functions in 'allelematch', reproduce the examples from the documentation, and include negative tests. The implementation is based on 'testthat'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/torstax/amregtest}
  \item Report bugs at \url{https://github.com/torstax/amregtest/issues}
}

}
\author{
\strong{Maintainer}: Torvald Staxler \email{torvald.staxler@telia.com}

Authors:
\itemize{
  \item Torvald Staxler \email{torvald.staxler@telia.com}
}

Other contributors:
\itemize{
  \item Department of Wildlife, Fish and Environmental Studies at Swedish University of Agricultural Sciences [copyright holder]
  \item Göran Spong \email{Goran.Spong@slu.se} (Senior Lecturer at the Department of Wildlife, Fish and Environmental Studies) [copyright holder]
  \item Paul Galpern \email{pgalpern@gmail.com} (Author of the excellent package 'allelematch' from which the five amExample data files have been copied under MIT license) [contributor]
}

}
\keyword{internal}
