% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amregtest-package.R
\docType{package}
\name{amregtest-package}
\alias{amregtest}
\alias{amregtest-package}
\title{amregtest: Runs Allelematch Regression Tests}
\description{
Automates regression testing when testing improvements of package '\link{allelematch}'.\cr Exported functions are prefixed with "art" for "'allelematch' regression test".\cr The exported functions are: \link{artVersion}, \link{artList} and \link{artRun}. \cr The tests are based on '\link{testthat}'. The tests use data sets as input. These data sets are stored within the package. See \link{artData} \cr The results of the tests are compared with pre-stored results, recorded with variants of \link{expect_snapshot}.
}
\author{
\strong{Maintainer}: Torvald Staxler \email{torvald.staxler@telia.com} [copyright holder]

}
\keyword{internal}
