% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amregtest-package.R
\docType{package}
\name{amregtest-package}
\alias{amregtest}
\alias{amregtest-package}
\title{amregtest: Runs allelematch regression tests}
\description{
This package is used to automate regression testing when testing improvments of package \link{allelematch}.\cr\cr Exported functions are prefixed with "art" for "allelematch regression test".\cr \cr The main functions are: \link{artRun} \link{artList} and \link{artVersion}. \cr\cr The tests are based on \link{testthat}. The tests use data sets as input. These data sets are stored within the package. See \link{amExampleData} and \link{ggSample} \cr\cr The results of the tests are compared with pre-stored results, recorded with variants of \link{expect_snapshot}.
}
\author{
\strong{Maintainer}: Torvald Staxler \email{torvald.staxler@telia.com}

}
\keyword{internal}
