% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amregtest-package.R
\name{amregtest}
\alias{amregtest}
\title{Package Overview}
\description{
Package 'amregtest' automates regression testing of package \link[allelematch:allelematch-package]{allelematch}.

The API is simple. There are only three functions:
\tabular{clcl}{
\verb{  } \tab \link{artRun}\verb{  }\tab \verb{  } \tab Executes the test, or a subset of the tests\cr
\verb{  } \tab \link{artList}\tab \tab Lists the available tests without running them\cr
\verb{  } \tab \link{artVersion}\tab \tab Shows the installed versions of 'allelematch' and 'amregtest'\cr
}

The prefix "art" is short for "Allelematch Regression Test".

See \link{artData} for a description of data sets used as input.
}
\references{
\link{amregtest-package}

\url{https://github.com/cran/allelematch}

\href{https://github.com/cran/allelematch/blob/2.5.1/inst/doc/allelematchSuppDoc.pdf}{allelematchSuppDoc.pdf}
}
