% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amregtest-package.R
\name{amregtest}
\alias{amregtest}
\title{Package Overview}
\description{
Package 'amregtest' automates regression testing of package \link[allelematch:allelematch-package]{allelematch}.

The API is simple. There are only four functions:
\tabular{clcl}{
\verb{  } \tab \code{\link{artRun}()}\verb{  }\tab \verb{  } \tab Executes the test, or a subset of the tests\cr
\verb{  } \tab \code{\link{artList}()}\tab \tab Lists the available tests without running them\cr
\verb{  } \tab \code{\link{artVersion}()}\tab \tab Shows the used versions of 'allelematch' and 'amregtest'\cr
\verb{  } \tab \code{\link{artInstallCranAllelematch}()}\tab\verb{  } \tab Installs official CRAN version of 'allelematch'\cr
}

The prefix "art" is short for "Allelematch Regression Test".

See \link{artData} for a description of data sets used as input.
}
\references{
\url{https://github.com/cran/allelematch}

\href{https://github.com/cran/allelematch/blob/2.5.1/inst/doc/allelematchSuppDoc.pdf}{allelematchSuppDoc.pdf}
}
