% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{amUniqueWrapper}
\alias{amUniqueWrapper}
\title{Wrapper around \link[allelematch:amUnique]{allelematch::amUnique}}
\usage{
amUniqueWrapper(amDatasetFocal, ...)
}
\arguments{
\item{amDatasetFocal}{A dataset from \link[allelematch:amDataset]{allelematch::amDataset}. Is passed on to \link[allelematch:amUnique]{allelematch::amUnique} as is.}

\item{...}{Logged and passed on to \link[allelematch:amUnique]{allelematch::amUnique} as is.}
}
\value{
The result from \link[allelematch:amUnique]{allelematch::amUnique}
}
\description{
Sets the locale to "C" and the language to "en" to ensure that
\link{sort} works the same way on all platforms. \cr
Generates a message string that describes the specified parameters.\cr
Feeds the specified parameters to \link[allelematch:amUnique]{allelematch::amUnique}.\cr
Measures the execution time.
}
