% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_data.R
\docType{data}
\name{artData}
\alias{artData}
\title{Example data used by \link{amregtest}}
\format{
Data frames with varying numbers of samples in rows, and alleles in columns. Missing data is represented as "-99".
}
\description{
This example data is used when testing allelematch backwards compatibility
using \link{artRun}. The tests load this data and passes it to \link{amDataset}.\cr
\cr
It includes data that was imported from version 5.2.1 of \link{allelematch}.
It was still unchanged in 5.2.4. \cr
\tabular{clcl}{
\verb{  } \tab \link{amExample1}\verb{  }\tab \verb{  } \tab Example 1 High quality data set\cr
\verb{  } \tab \link{amExample2}\tab \tab Example 2 Good quality data set\cr
\verb{  } \tab \link{amExample3}\tab \tab Example 3 Marginal quality data set\cr
\verb{  } \tab \link{amExample4}\tab \tab Example 4 Low quality data set\cr
\verb{  } \tab \link{amExample5}\tab \tab Example 5 Wildlife data set\cr
}
See \href{https://github.com/cran/allelematch/blob/2.5.1/inst/doc/allelematchSuppDoc.pdf}{allelematchSuppDoc.pdf}
for a more detailed description. \cr

It also includes a large data set gathered from field work:\cr
\tabular{clcl}{
\verb{  } \tab \link{ggSample}\verb{  }\tab \verb{  } \tab Very large wildlife data set\cr
}
}
\references{
\url{https://github.com/cran/allelematch}

\href{https://github.com/cran/allelematch/blob/2.5.1/inst/doc/allelematchSuppDoc.pdf}{allelematchSuppDoc.pdf}
}
\keyword{data}
