% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_api.R
\name{artInstallCranAllelematch}
\alias{artInstallCranAllelematch}
\title{Installs official version of 'allelematch' from CRAN.}
\usage{
artInstallCranAllelematch(version = "2.5.5")
}
\arguments{
\item{version}{string. Default "2.5.5".}
}
\value{
TBD
}
\description{
Note that the CRAN version will be built from source code at installation.
This means that \link{artVersion} will show a "(Built HH:MM)" timestamp
from the installation rather than from when it was published on CRAN.
}
\examples{
# Install the default official version of 'allelematch' from CRAN:
artInstallCranAllelematch()

# Install another official version of 'allelematch' from CRAN:
artInstallCranAllelematch("2.5.3")

}
\seealso{
\link{artVersion}, \link{artList}, \link{artRun} and \link{amregtest}
}
