% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artList}
\alias{artList}
\title{Lists available tests in \code{amregtest} without running them}
\usage{
artList()
}
\value{
A vector containing all the tests names
}
\description{
Use the output to select a value for parameter \code{filter} to \link{artRun}.
Useful when debugging.
}
\examples{
# List the available tests:
artList()
\dontrun{

# Run the first of the available tests:
artRun(filter="allelematch_1-amDataset")

# Run all tests that start with 'am' and 'allelematch_1-amDataset':
artRun(filter="^am|allelematch_1-amDataset")
}

}
\seealso{
\link{artRun}
}
