% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artRun}
\alias{artRun}
\title{Runs the regressiontest}
\usage{
artRun()
}
\value{
NA
}
\description{
TODO:
}
\seealso{
\link{artVersion} \link{artAssertAllelematchVersion}

\link{artSet} \link{artShow} \link{artClear}
}
