% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artRun}
\alias{artRun}
\title{Runs the regression test}
\usage{
artRun(skip_slow = FALSE, generate_html = FALSE)
}
\arguments{
\item{skip_slow}{TRUE or FALSE}

\item{generate_html}{TRUE or FALSE}
}
\value{
NA
}
\description{
TODO:
}
\seealso{
\link{artVersion} \link{artAssertAllelematchVersion}

\link{artSet} \link{artShow} \link{artClear}
}
