% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_api.R
\name{artVersion}
\alias{artVersion}
\title{Prints package version}
\usage{
artVersion()
}
\description{
Prints version of this package (\link{amregtest}) and of tested version of \link{allelematch}.\cr
\cr
The version is specified in the file DESCRIPTION, tag "Version: ".
}
\examples{
# See what version of packages 'allelematch' and 'amregtest'
# are currently installed:
artVersion()

# List the available tests:
artList()

\dontrun{

# Run all the tests:
artRun()

# Run the first of the available tests:
artRun(filter="allelematch_1-amDataset")

# Run all tests that start with 'am' and 'allelematch_1-amDataset':
artRun(filter="^am|allelematch_1-amDataset")

# Run the test 'amExample3' and let it generate .html files
# in the [getwd] directory:
artRun(filter='amExample3', html=TRUE)
}


}
\seealso{
\link{artList}, \link{artRun} and \link{amregtest}
}
