% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artVersion}
\alias{artVersion}
\title{Prints package version}
\usage{
artVersion()
}
\description{
Prints version of this package (\link{amregtest}) and of tested version of \link{allelematch}.\cr
\cr
The version is specified in the file DESCRIPTION, tag "Version: ".
}
\seealso{
\link{artRun} \link{artList}
}
