% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{artVersion}
\alias{artVersion}
\title{Prints package version}
\usage{
artVersion()
}
\description{
Prints version of this package and of tested version of \link{allelematch}.
}
\seealso{
\link{artInstallAllelematchVersion}
}
