% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_text.R
\name{auAssertCsvEqualToExpectedData}
\alias{auAssertCsvEqualToExpectedData}
\title{Verifies that the data in \code{actualCsvFile} exactly matches that in \code{expectedData}}
\usage{
auAssertCsvEqualToExpectedData(actualCsvFile, expectedData = NULL)
}
\description{
Fetches expected from the \link{regressiontest} data using \link[utils:data]{utils::data} \cr
\cr
Note that \link{allelematch} and \link[utils:data]{utils::data} differs on the how .csv files
shall be encoded.\cr
o \link{allelematch} writes .csv files with comma (",") as separator.\cr
o \link{data} expects .csv files to have semicolon (";") as separator.
}
