% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_alparam.R
\name{auDefaultFileName}
\alias{auDefaultFileName}
\title{Turns \link{amUniqueParamString} into a string that can be used in a filename}
\usage{
auDefaultFileName(...)
}
\arguments{
\item{...}{}
}
\value{
a string that can be used in a data filename for use in an allelematch regressiontest.
}
\description{
Turns \link{amUniqueParamString} into a string that can be used in a filename
}
