% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_misc.R
\name{getdata}
\alias{getdata}
\title{Returns a large data set from some package without duplicating}
\usage{
getdata(...)
}
\description{
The parameters are the same as for the \code{utils::data()} function.
}
\details{
Thank you @henfiber! See https://stackoverflow.com/questions/30951204/load-dataset-from-r-package-using-data-assign-it-directly-to-a-variable/30951700#30951700
}
