% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_data.R
\name{getdata}
\alias{getdata}
\title{Loads and returns a large data set without polluting the global environment.}
\usage{
getdata(name, ...)
}
\arguments{
\item{name}{The name (excluding directory and extension) of the data to load}

\item{...}{All other parameters are passed on to \link[utils:data]{utils::data} as is}
}
\value{
the specified data set loaded by \link[utils:data]{utils::data}
}
\description{
The data set is taken from the ./data/ folder of the specified package.
}
\details{
See the Description section under \link[utils:data]{utils::data} for how the supported file
types (including .R, .RData, .txt and .csv) shall be encoded.

The parameters are the same as for the \link{data} function in package \link{utils}.

Thank you @henfiber! See https://stackoverflow.com/questions/30951204/load-dataset-from-r-package-using-data-assign-it-directly-to-a-variable/30951700#30951700
}
\examples{
foo = getdata("amExample1", package = "allelematch")

}
