% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_testengin.R
\name{testDataSet}
\alias{testDataSet}
\title{Runs a set of tests on the supplied dataset}
\usage{
testDataSet(dataSetDir = "data/", skippCleaningInput = FALSE)
}
\arguments{
\item{dataSetDir}{The directory that contains the data sets to be tested. (string)}

\item{skippCleaningInput}{If FALSE, the input files are cleaned up
and written to a new file that adds "_clean" to the name of the original files.
If TRUE, this function skips reading the original files and read the cleaned
files directly instead. (boolean)}
}
\value{
TODO: Change to return TRUE on success and FALSE on failure.
}
\description{
\code{testDataSet} reads allelematch test input files from and writes output files
to the \code{dataSetDir}.
}
\details{
\code{testDataSet} is the entry point for the \code{regressiontest} package.

Both input and output data files in \code{dataSetDir} have predefined names.

The input files are called "input_new_samples.txt" and "input_Match_references.txt".
The input files are read into data frames at the start of \code{testDataSet}.
The same input data is used in all calls to the \code{allelematch} functions
to be tested.

The output files have names that describe the called \code{allelematch} functions
and the parameters that are passed to the same functions.
}
